# Adding replication benchmark