I created a subproject to hold the benchmarks and added one for replication. To avoid too many changes in this commit, I'll move the existing benchmarks later.

This benchmark is not really micro. We utilize the `RaftHandle` instances to replicate the data with Raft.
We measure the throughput by varying the cluster and data sizes.
This is similar to how a user configures and uses the instances.

The goal is to add a benchmark to start working towards JDK 21 features.
Especially around the virtual threads, so we want to measure the performance with enabled/disabled.
And at later stages, we could measure when applying some architectural changes.

@franz1981, let me know what you think when you're back.