@jabolina 2024-01-20
I am running this benchmark on JDK 21 (21+35) on an AMD Ryzen 5 3500X 6-Core Processor.

First, a baseline with `InMemoryLog` implementation. We have no writes to disk with this one.

![in-memory-log](https://github.com/jgroups-extras/jgroups-raft/assets/13581903/c1a135be-90d6-4a55-ba70-14ee8d41b9ec)

And the JMH output for that was:

```text
Benchmark                                 (clusterSize)  (dataSize)   Mode  Cnt       Score      Error  Units
DataReplicationBenchmark.testReplication              1         256  thrpt    5  156554.879 ± 6854.907  ops/s
DataReplicationBenchmark.testReplication              1        1024  thrpt    5  108758.445 ±  746.293  ops/s
DataReplicationBenchmark.testReplication              1        4096  thrpt    5   28750.762 ±   16.965  ops/s
DataReplicationBenchmark.testReplication              3         256  thrpt    5   21411.663 ± 1395.440  ops/s
DataReplicationBenchmark.testReplication              3        1024  thrpt    5   21466.775 ± 1536.580  ops/s
DataReplicationBenchmark.testReplication              3        4096  thrpt    5   22084.768 ±  735.159  ops/s
DataReplicationBenchmark.testReplication              5         256  thrpt    5   19412.594 ±  483.408  ops/s
DataReplicationBenchmark.testReplication              5        1024  thrpt    5   18729.931 ±  432.269  ops/s
DataReplicationBenchmark.testReplication              5        4096  thrpt    5   15116.383 ±  696.190  ops/s
```

The thing here is the drop in the single node going from 1024 bytes to 4096.
I did a quick check, and the culprit here seems to be the message fragmentation.
We keep waiting on the `CreditMap` for credits to be available.
I can get some prints, if necessary. But should be easy to check that.

## FileBasedLog

Now, starting with stable storage in the system.
I ran the benchmark enabling and disabling fsync.

### fsync disabled

Starting with fsync disabled.

![file-based-log](https://github.com/jgroups-extras/jgroups-raft/assets/13581903/ed1a53f8-3ea3-41c6-9594-466168a3206f)

We have the same drop as the in-memory log.
And, in the cluster, we have a similar throughput to in-memory, too.
I guess the replication to other nodes dominates everything.

The JMH output here was:

```text
Benchmark                                 (clusterSize)  (dataSize)   Mode  Cnt       Score      Error  Units
DataReplicationBenchmark.testReplication              1         256  thrpt    5  142746.620 ± 5649.177  ops/s
DataReplicationBenchmark.testReplication              1        1024  thrpt    5  108998.248 ±  300.977  ops/s
DataReplicationBenchmark.testReplication              1        4096  thrpt    5   28760.314 ±   36.031  ops/s
DataReplicationBenchmark.testReplication              3         256  thrpt    5   21581.735 ±  914.821  ops/s
DataReplicationBenchmark.testReplication              3        1024  thrpt    5   20477.593 ± 2178.954  ops/s
DataReplicationBenchmark.testReplication              3        4096  thrpt    5   20432.116 ±  702.711  ops/s
DataReplicationBenchmark.testReplication              5         256  thrpt    5   18825.619 ±  173.113  ops/s
DataReplicationBenchmark.testReplication              5        1024  thrpt    5   17729.728 ±  122.396  ops/s
DataReplicationBenchmark.testReplication              5        4096  thrpt    5   14518.041 ±  815.735  ops/s
```

### fsync enabled

With fsync enabled.

![file-based-log-fsync](https://github.com/jgroups-extras/jgroups-raft/assets/13581903/c5e4e0a5-930e-4816-add1-cdda88641992)

We have a drop on the case for single-node and a smaller message.
Other than that, everything is pretty similar.
The JMH output for this run:

```text
Benchmark                                 (clusterSize)  (dataSize)   Mode  Cnt       Score      Error  Units
DataReplicationBenchmark.testReplication              1         256  thrpt    5  140909.868 ± 5541.895  ops/s
DataReplicationBenchmark.testReplication              1        1024  thrpt    5  108922.101 ±  917.301  ops/s
DataReplicationBenchmark.testReplication              1        4096  thrpt    5   28765.462 ± 1847.612  ops/s
DataReplicationBenchmark.testReplication              3         256  thrpt    5   21694.445 ±  835.020  ops/s
DataReplicationBenchmark.testReplication              3        1024  thrpt    5   20986.808 ±  904.307  ops/s
DataReplicationBenchmark.testReplication              3        4096  thrpt    5   20017.131 ±  573.814  ops/s
DataReplicationBenchmark.testReplication              5         256  thrpt    5   18212.267 ±  255.478  ops/s
DataReplicationBenchmark.testReplication              5        1024  thrpt    5   17580.337 ±  818.260  ops/s
DataReplicationBenchmark.testReplication              5        4096  thrpt    5   14580.382 ±  321.329  ops/s
```

## LevelDB

Now, the executions utilize the `LevelDBLog` implementation.
The default configuration.

Bela mentioned that enabling fsync with LevelDB doesn't actually flush everything.
So the comparisons with fsync could be unfair.

### fsync disabled

With fsync set to false.

![leveldb-log](https://github.com/jgroups-extras/jgroups-raft/assets/13581903/880e7633-4133-400e-8a08-56c9e3a76db1)

Not to brag, but the log implementation we ship is performing better here.
In addition, the confidence intervals are a bit larger than the others.
I ran that a few times and the results were all similar.
The JMH output:

```text
Benchmark                                 (clusterSize)  (dataSize)   Mode  Cnt      Score       Error  Units
DataReplicationBenchmark.testReplication              1         256  thrpt    5  93523.196 ± 11632.784  ops/s
DataReplicationBenchmark.testReplication              1        1024  thrpt    5  81565.789 ±  4244.783  ops/s
DataReplicationBenchmark.testReplication              1        4096  thrpt    5  28750.379 ±  1266.126  ops/s
DataReplicationBenchmark.testReplication              3         256  thrpt    5  19237.459 ±  1767.513  ops/s
DataReplicationBenchmark.testReplication              3        1024  thrpt    5  17675.199 ±  2231.221  ops/s
DataReplicationBenchmark.testReplication              3        4096  thrpt    5  14085.667 ±  1308.144  ops/s
DataReplicationBenchmark.testReplication              5         256  thrpt    5  16251.385 ±  1335.671  ops/s
DataReplicationBenchmark.testReplication              5        1024  thrpt    5  14477.722 ±   552.461  ops/s
DataReplicationBenchmark.testReplication              5        4096  thrpt    5   9424.252 ±  1622.021  ops/s
```

### fsync enabled

With fsync set to true.

![leveldb-log-fsync](https://github.com/jgroups-extras/jgroups-raft/assets/13581903/4c070d61-b59b-40db-a4d5-5e9a11c4d28a)

Quite similar to the previous run with fsync disabled.
And still, our file-based log shows a better throughput here.