@belaban 2024-01-22
Why do you measure 1 node? I don't think we've optimized for this, but then again, this use case is (IMO) only academic...