@franz1981 2024-01-22
@jabolina looking at the numbers, they look weird....using or not using fsync don't make any difference, for the file log, which means...
1. fsync is not actually tuned (I've added this https://github.com/jgroups-extras/jgroups-raft/pull/244#discussion_r1461061407 comment for that)
2. the machine used doesn't make it happen
3. the bottleneck isn't there