@jabolina 2024-01-26
With the cache disabled, I had no noticeable difference with and without fsync. :frowning_face: 

I run the [LogJmhBenchmark](https://github.com/jgroups-extras/jgroups-raft/blob/master/tests/junit-functional/org/jgroups/perf/LogJmhBenchmark.java), with parameters `-plogType="file" -pbaseDir="/home/jab/raft-log-bench" -pbatchSize=1`. I made some changes to the output in milliseconds, and the dataSize the same as the new benchmark in this PR (256, 1024, 4096):

Without fsync:
```text
Benchmark                              (baseDir)  (batchSize)  (dataSize)  (logType)  Mode  Cnt  Score   Error  Units
LogJmhBenchmark.append  /home/jab/raft-log-bench            1         256       file  avgt   10  0.003 ± 0.002  ms/op
LogJmhBenchmark.append  /home/jab/raft-log-bench            1        1024       file  avgt   10  0.009 ± 0.002  ms/op
LogJmhBenchmark.append  /home/jab/raft-log-bench            1        4096       file  avgt   10  0.041 ± 0.038  ms/op
```

With fsync:
```text
Benchmark                              (baseDir)  (batchSize)  (dataSize)  (logType)  Mode  Cnt   Score   Error  Units
LogJmhBenchmark.append  /home/jab/raft-log-bench            1         256       file  avgt   10   2.293 ± 0.015  ms/op
LogJmhBenchmark.append  /home/jab/raft-log-bench            1        1024       file  avgt   10   5.758 ± 2.162  ms/op
LogJmhBenchmark.append  /home/jab/raft-log-bench            1        4096       file  avgt   10  10.749 ± 0.305  ms/op
```

Quite a difference. In the benchmark here, the bottleneck might be somewhere else. Does what we have here serve as a baseline for starting work? The worrying part is that it covers too much, and it would be difficult to identify if optimizations are working. :worried: 

And just for curiosity, running the `LogJmhBenchmark` with fsync but with the nvme cache enabled:

```text
Benchmark                              (baseDir)  (batchSize)  (dataSize)  (logType)  Mode  Cnt  Score   Error  Units
LogJmhBenchmark.append  /home/jab/raft-log-bench            1         256       file  avgt   10  1.319 ± 1.062  ms/op
LogJmhBenchmark.append  /home/jab/raft-log-bench            1        1024       file  avgt   10  4.022 ± 0.161  ms/op
LogJmhBenchmark.append  /home/jab/raft-log-bench            1        4096       file  avgt   10  6.108 ± 0.248  ms/op
```